fix: adds exception to clippy::unused_unit warnings.
`wasm_bindgen` sometimes generates an orphaned unit expression
as part of its macro expansion. This is relatively harmless, but
is flagged by clippy, this codebase's choice for Rust linter.

We have no real control on how `#[wasm_bindgen]` expands,
and no real way to avoid this from happening in the expanded code,
so we opt to disable the `clippy::unused_unit` lint check
on affected files.

Definitely not ideal, but spotting unused units on code we're
writing ourselves are relatively easy to spot anyway.

@see: https://github.com/rustwasm/wasm-bindgen/issues/2774